Add httpretty to functests
We don't want our functional tests to be talking to the real network
(e.g. Google's OAuth 2 server), for the same reason that we don't want
our unit tests to.
