+++ requirements/functests.in
@@ -1,3 +1,4 @@
+httpretty
 pytest
 webtest
 h-matchers
