+++ tests/conftest.py
@@ -1,5 +1,6 @@
 import os
 
+import httpretty
 import pytest
 
 from checkmate.db import create_engine
@@ -30,3 +31,19 @@ def pyramid_settings():
         "public_scheme": "http",
         "public_port": "9099",
     }
+
+
+@pytest.fixture(autouse=True)
+def httpretty_():
+    """Monkey-patch Python's socket core module to mock all HTTP responses.
+
+    We never want real HTTP requests to be sent by the tests so replace them
+    all with mock responses. This handles requests sent using the standard
+    urllib2 library and the third-party httplib2 and requests libraries.
+    """
+    httpretty.enable(allow_net_connect=False)
+
+    yield
+
+    httpretty.disable()
+    httpretty.reset()
