+++ tests/unit/conftest.py
@@ -4,7 +4,6 @@ import functools
 from unittest.mock import MagicMock
 from urllib.parse import urlencode
 
-import httpretty
 import sqlalchemy
 from pyramid import testing
 from pyramid.request import Request, apply_request_extensions
@@ -102,19 +101,3 @@ def db_session(db_engine):
         session.close()  # pylint:disable=no-member
         trans.rollback()
         conn.close()
-
-
-@pytest.fixture(autouse=True)
-def httpretty_():
-    """Monkey-patch Python's socket core module to mock all HTTP responses.
-
-    We never want real HTTP requests to be sent by the tests so replace them
-    all with mock responses. This handles requests sent using the standard
-    urllib2 library and the third-party httplib2 and requests libraries.
-    """
-    httpretty.enable(allow_net_connect=False)
-
-    yield
-
-    httpretty.disable()
-    httpretty.reset()
